+++ src/view/elements/gxViewElement.h
@@ -3,8 +3,13 @@
 
 #include "core/geometry/gxGeometry.h"
 #include "core/gxComposite.h"
+#include "core/gxFlags.h"
 #include "view/gxPainter.h"
 #include "view/gxBounds.h"
+
+
+// char used as 1 byte int
+#define gxViewElementFlags char
 
 // Forward declaration
 class gxRootViewElement;
@@ -84,8 +89,8 @@ protected:
   virtual void PaintChildren(gxPainter &aPainter) = 0;
 
   /**
-   * @brief Returns the {@link gxRootViewElement root element} of the hierarchy by
-   * traversing up the hierarchy until hitting the root.
+   * @brief Returns the {@link gxRootViewElement root element} of the hierarchy
+   * by traversing up the hierarchy until hitting the root.
    */
   virtual gxRootViewElement* GetRootViewElement();
 
@@ -93,11 +98,58 @@ protected:
    * @brief Repaints the view element making its bounds dirty.
    */
   virtual void Repaint();
+  
   /**
    * @brief Repaints part or the whole of the view element.
    * @param aBounds The bounds of the area to be repainted.
    */
   virtual void Repaint(gxBounds &aBounds);
+
+  // Validation related methods
+  
+  /**
+   * @brief Marks the view element is invalid.
+   */
+  virtual void Invalidate();
+
+  /**
+   * @brief Marks the view element as invalid and calls Revalidate on all
+   * parents.
+   * 
+   * This method is called whenever the size or position of the element change,
+   * so to let parent objects a change to respond to the change.
+   * 
+   * For example, a layout will need to update all children if one changes
+   * position; a Scroller might need to readjust the scrollbar range when a
+   * child grows in size.
+   */
+  virtual void Revalidate();
+
+
+  /**
+   * @brief Calls ValidateSelf(), and then Validate() on all children.
+   */
+  virtual void Validate();
+
+  /**
+   * @brief Asks the view element to perform any operations that are needed
+   * to make it valid
+   */
+  virtual void ValidateSelf() { }
+  
+  /**
+   * @brief Returns whether or not the view element is valid.
+   * @return True if the view element is valid.
+   */
+  bool IsValid();
+
+  enum Flags
+  {
+    Valid = 0x01
+  };
+  
+  gxFlags<gx8Flags> mFlags;
+
 
   IMPLEMENT_COMPOSITE(gxViewElement)
 };