+++ changelog.md
@@ -1,5 +1,6 @@
 v0.1.03
 -------
+- Added: Validation mechanism.
 - Added: Scroll support.
 - Added: gxObserverList as the gefri's core subject-observer mechanism.
 - Added: gxCallback - a non-static member method callback wrapper (for gxOberverList).